myenv/requirements.txt
flask
langchain
pinecone-client
openai
tiktoken
python-dotenv
pymongo
boto3
flask-cors